requirements.txt
fastapi==0.123.0
pydantic==2.12.5
SQLAlchemy==2.0.44
python-dotenv==1.2.1
python-jose==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.3.0
uvicorn[standard]==0.38.0
